requirements.txt
markdown
pygments
PyQt6
PyQt6-WebEngine
pynput
pyperclip
comtypes
pillow
yaml
pygetwindow
openai